requirements.txt
openai-whisper==20231117
torch>=2.0.0
numpy==1.24.3
ffmpeg-python>=0.2.0
setuptools>=65.5.1
tqdm>=4.65.0
more-itertools>=10.1.0
tiktoken>=0.3.3
numba>=0.57.1
langchain>=0.1.0
langchain-community>=0.0.13
langchain-huggingface>=0.0.6
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
openai>=1.12.0
chromadb==0.4.24
gTTS>=2.3.2
pygame>=2.5.1
pyttsx3>=2.90 